rename repo-jest-setup to repo-test-setup
