+++ vitest.config.mjs
@@ -20,7 +20,7 @@ const EXCLUDE_FROM_TESTING = [
 const EXCLUDE_FROM_COVERAGE = [
   ...EXCLUDE_FROM_TESTING,
   'src/**/*.test.*',
-  'repo-jest-setup.jsx',
+  'repo-test-setup.jsx',
   'vitest.setup.ts',
   'custom-testing-library.js',
   'setupTestGlobal.js',
